+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "package-json-upgrade" extension will be documented i
 
 ## Unreleased changes
 
+- Add config for minimumReleaseAge/minimumReleaseAgeExclude, mirroring the pnpm setting
 - Display better error messages when failing to find dependency/version
 
 ## 3.6.1